Split merch.test.ts into merch.store.test.ts and merch.order.test.ts
